Completed Routes
Starting now on completing the Models to clean up code
